refactor: restructure package for v2 with slog as default
- Update module path to github.com/gobuffalo/logger/v2
- Move slog implementation to unexported defaultFieldLogger in default.go
- Separate interfaces into logger.go (Logger, FieldLogger, Outable)
- Rename logrus.Logrus to unexported fieldLogger for consistency
- Update logrus dependency to v1.9.4
- Remove outable.go and slog.go (consolidated into default.go)

BREAKING CHANGE: Module path now includes /v2 suffix
