+++ go.mod
@@ -1,10 +1,10 @@
-module github.com/gobuffalo/logger
+module github.com/gobuffalo/logger/v2
 
 go 1.25
 
 require (
-	github.com/sirupsen/logrus v1.9.1
+	github.com/sirupsen/logrus v1.9.4
 	golang.org/x/term v0.0.0-20220722155259-a9ba230a4035
 )
 
-require golang.org/x/sys v0.0.0-20220722155257-8c9f86f7a55f // indirect
+require golang.org/x/sys v0.13.0 // indirect
