+++ logrus/logrus.go
@@ -6,55 +6,55 @@
 //
 // Usage:
 //
-//	import "github.com/gobuffalo/logger/logrus"
+//	import "github.com/gobuffalo/logger/v2/logrus"
 //	log := logrus.New(logrus.InfoLevel)
 //	log.Info("Application started")
 //	log.WithField("version", "1.0.0").Info("Version info")
 //
-// The Logrus type implements the logger.FieldLogger interface, making it
+// The fieldLogger type implements the logger.FieldLogger interface, making it
 // interchangeable with the default logger implementation.
 package logrus
 
 import (
 	"io"
 
-	"github.com/gobuffalo/logger"
+	"github.com/gobuffalo/logger/v2"
 	"github.com/sirupsen/logrus"
 )
 
-var _ logger.Logger = Logrus{}
-var _ logger.FieldLogger = Logrus{}
-var _ logger.Outable = Logrus{}
+var _ logger.Logger = fieldLogger{}
+var _ logger.FieldLogger = fieldLogger{}
+var _ logger.Outable = fieldLogger{}
 
-// Logrus is a Logger implementation backed by sirupsen/logrus
-type Logrus struct {
+// fieldLogger is a Logger implementation backed by sirupsen/logrus
+type fieldLogger struct {
 	logrus.FieldLogger
 }
 
 // SetOutput will try and set the output of the underlying
 // logrus.FieldLogger if it can
-func (l Logrus) SetOutput(w io.Writer) {
+func (l fieldLogger) SetOutput(w io.Writer) {
 	if lg, ok := l.FieldLogger.(interface{ SetOutput(io.Writer) }); ok {
 		lg.SetOutput(w)
 	}
 }
 
-func (l Logrus) WithField(s string, i any) logger.FieldLogger {
-	return Logrus{l.FieldLogger.WithField(s, i)}
+func (l fieldLogger) WithField(s string, i any) logger.FieldLogger {
+	return fieldLogger{l.FieldLogger.WithField(s, i)}
 }
 
-func (l Logrus) WithFields(m map[string]any) logger.FieldLogger {
-	return Logrus{l.FieldLogger.WithFields(logrus.Fields(m))}
+func (l fieldLogger) WithFields(m map[string]any) logger.FieldLogger {
+	return fieldLogger{l.FieldLogger.WithFields(logrus.Fields(m))}
 }
 
-// New creates a new Logrus logger.
+// New creates a new logger.FieldLogger backed by logrus.
 func New(lvl logger.Level) logger.FieldLogger {
 	l := logrus.New()
 	l.SetLevel(logrus.Level(lvl))
-	return Logrus{l}
+	return fieldLogger{l}
 }
 
-// NewLogger creates a new Logrus logger from a string level.
+// NewLogger creates a new logger.FieldLogger from a string level.
 func NewLogger(level string) logger.FieldLogger {
 	lvl, err := logrus.ParseLevel(level)
 	if err != nil {
