--- outable.go
@@ -1,8 +1,0 @@
-package logger
-
-import "io"
-
-// Outable interface for loggers that allow setting the output writer
-type Outable interface {
-	SetOutput(out io.Writer)
-}
